+++ R/sensor_filterMeta.R
@@ -21,13 +21,22 @@
 #' @seealso \link{sensor_filter}
 #' @seealso \link{sensor_filterDate}
 #' 
-#' @examples \donttest{
+#' @examples
 #' library(AirSensor)
 #' 
-#' example_sensor %>%
-#'   sensor_filterMeta(stateCode == "CA") %>%
-#'   PWFSLSmoke::monitor_stamenmap(zoom = 11)
-#' }
+#' example_sensor_scaqmd %>%
+#'   sensor_extractMeta() %>%
+#'   dplyr::pull("communityRegion") %>%
+#'   sort () %>%
+#'   unique()
+#'   
+#' example_sensor_scaqmd %>%
+#'   sensor_filterMeta(communityRegion == "Imperial Valley") %>%
+#'   sensor_extractMeta() %>%
+#'   dplyr::pull("communityRegion") %>%
+#'   sort () %>%
+#'   unique()
+#'
 
 sensor_filterMeta <- function(
   sensor = NULL,